23/10/25
Contribution guidelines made from contributing.md and The Turing Way repurposed for this project.

FAQ file has outline of what is needed to populate it.

things_to_think about page has a skeleton of what it should eventually contain. A sort of checklist for those looking to start a coding CAFE